Code improvements
Improved readability of `goalRow`, `matsRow`, and `remRow` functions in `CharacterGoalTable`, `RosterGoalTable`, `MatsTable`, and `RemTable`.

`remRow` now returns an Element component instead of an array of elements, which simplifies the code in `RemTable`. However, I was unable to do this for `goalRow` or `matsRow`, as it seems that doing so causes child inputs to lose focus each time the component is re-rendered, which greatly harms the user experience. `RemTable` is always read-only, so the problem does not apply to `RemTable`.