RosterStorageCard colors implementation
Renamed `RosterStorageTable` to `RosterStorageCard` to be more consistent with the naming scheme of other components (where Table is generally a component nested within a Card, and Card is generally a component nested within the top level of the tab).

Implemented colors for `RosterStorageCard` as well as dual colors for combo `RosterStorageCard`s.