RosterStorageCard handleChange bugfix
`RosterStorageCard` had a bug in `handleChange` where total quantity of a material would be erroneously updated even if the material was not selected for use. Total quantity is now only updated if the material is selected for use.
Note: I mentioned selection chests in the previous commit's TODO, but it turns out this bug was not necessarily tied to selection chests.