Added daily chest sync to `RosterStorageCard`
Added the ability for two separate `RosterStorageCard`s to synchronize daily chest state. This is because one of the daily chests is a selection chest between leaps and shards, which are otherwise generally separate materials (hence the choice to not use a combo table with selectionChest). A separate combo table also would have required synced state for the material totals, so I decided this was the best approach.

Also changed default selected to `true` for most sources, as I think it is a more intuitive experience after testing.