Roster storage quantity field improvement
`SourceRow` now uses a controlled input for quantity, which allows updating either quantity field for a selection chest. This leads to a better UX because updating quantity on the side with higher "use" value set will favor that higher "use" value instead of essentially resetting the user's previous "use" value.