Improved formatting on vertical screens
`MarketDataTable` does not stretch the entire screen if the screen is too small horizontally for it to be side by side with roster storage. `RosterStorageTable` can now render wider columns for Quantity and Amount if configured to; this is helpful for displaying large values of silver, gold, shards, etc. on vertical screens.