Sitewide welcome message modal
Implemented a sitewide welcome message modal that displays on first visit, or if a version number update occurred since the last visit. This welcome message displays an overview of LOA Planner features with a goal of guiding users to use them effectively.

Also added Goal Milestones to the Planned Features page.