Quality of life upgrades
External shortcut button for Maxroll Upgrade Calculator added to Roster View (opens in new tab).

Character limit expanded, as the amount of data is less of a concern with improvements to saving in `character-data.ts`. It's probably still worth having a character limit to avoid edge cases of flooding the UI in the "Configure Roster Goals" modal or approaching the limit of local storage size.